Refactor DragAndDropFilesWidget to separate file for structure.
